# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenAgents API",
        "description": "The OpenAgents API",
        "contact": {
            "name": "OpenAgents",
            "url": "https://x.com/OpenAgentsInc"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://openagents.com/api/v1",
            "description": "OpenAgents API server"
        }
    ],
    "paths": {
...
        "/agents/{agentId}/files": {
            "post": {
                "tags": [
                    "AgentFile"
                ],
                "summary": "Create agent file",
                "operationId": "storeAgentFile",
                "parameters": [
                    {
                        "name": "agentId",
                        "in": "path",
                        "description": "ID of the agent to add file for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload new file for agent",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "file_id": {
                                                    "type": "integer"
                                                },
                                                "agent_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}